small clean up and refactoring
